+++ NEWS.md
@@ -12,6 +12,9 @@ Changelog
   + default value is `2021-01-17` (YYYY-mm-dd or %Y-%-m-%d)
   + value read from config_db `Settings` table, with the `setting` value of `dateformat`
   + associated choices `self$dateformat_choices`, a vector of character strings
+* new `self$dateformat_choice` active variable.
+  + reads current value of `self$dateformat` OR
+  + writes to `self$dateformat` and writes to config_db `Settings`
 
 # 0.9.15
 16th June 2021